feat 判断歌曲是否已加入列表
