+++ app/renderer/src/utils/index.ts
@@ -1,4 +1,8 @@
-export { default as HowlPlayer, SinglePlayer } from './player';
+export {
+  default as HowlPlayer,
+  SinglePlayer,
+  type PlayerOrder,
+} from './player';
 
 export const randomInt = (m: number, n: number) =>
   Math.floor(Math.random() * (m - n + 1)) + n;
